feat: Add reverse search capability for shell commands (#4793)
