+++ packages/cli/src/ui/hooks/useSlashCompletion.test.ts
@@ -8,7 +8,7 @@
 
 import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
-import { useCompletion } from './useCompletion.js';
+import { useSlashCompletion } from './useSlashCompletion.js';
 import * as fs from 'fs/promises';
 import * as path from 'path';
 import * as os from 'os';
@@ -16,7 +16,7 @@ import { CommandContext, SlashCommand } from '../commands/types.js';
 import { Config, FileDiscoveryService } from '@google/gemini-cli-core';
 import { useTextBuffer } from '../components/shared/text-buffer.js';
 
-describe('useCompletion', () => {
+describe('useSlashCompletion', () => {
   let testRootDir: string;
   let mockConfig: Config;
 
@@ -50,7 +50,7 @@ describe('useCompletion', () => {
 
   beforeEach(async () => {
     testRootDir = await fs.mkdtemp(
-      path.join(os.tmpdir(), 'completion-unit-test-'),
+      path.join(os.tmpdir(), 'slash-completion-unit-test-'),
     );
     testDirs = [testRootDir];
     mockConfig = {
@@ -82,12 +82,13 @@ describe('useCompletion', () => {
           { name: 'dummy', description: 'dummy' },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest(''),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
             mockConfig,
           ),
         );
@@ -112,12 +113,13 @@ describe('useCompletion', () => {
         const { result, rerender } = renderHook(
           ({ text }) => {
             const textBuffer = useTextBufferForTest(text);
-            return useCompletion(
+            return useSlashCompletion(
               textBuffer,
               testDirs,
               testRootDir,
               slashCommands,
               mockCommandContext,
+              false,
               mockConfig,
             );
           },
@@ -143,12 +145,13 @@ describe('useCompletion', () => {
         ] as unknown as SlashCommand[];
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/help'),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
             mockConfig,
           ),
         );
@@ -176,12 +179,13 @@ describe('useCompletion', () => {
           { name: 'dummy', description: 'dummy' },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest(''),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
             mockConfig,
           ),
         );
@@ -198,12 +202,14 @@ describe('useCompletion', () => {
           { name: 'dummy', description: 'dummy' },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest(''),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -223,12 +229,14 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/h'),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -251,12 +259,14 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/h'),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -280,12 +290,14 @@ describe('useCompletion', () => {
           { name: 'chat', description: 'Manage chat' },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/'),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -326,12 +338,14 @@ describe('useCompletion', () => {
         })) as unknown as SlashCommand[];
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/command'),
             testDirs,
             testRootDir,
             largeMockCommands,
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -384,7 +398,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/'),
             testDirs,
             testRootDir,
@@ -407,7 +421,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/mem'),
             testDirs,
             testRootDir,
@@ -431,7 +445,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/usag'), // part of the word "usage"
             testDirs,
             testRootDir,
@@ -458,7 +472,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/clear'), // No trailing space
             testDirs,
             testRootDir,
@@ -490,7 +504,7 @@ describe('useCompletion', () => {
           ] as unknown as SlashCommand[];
 
           const { result } = renderHook(() =>
-            useCompletion(
+            useSlashCompletion(
               useTextBufferForTest(query),
               testDirs,
               testRootDir,
@@ -511,7 +525,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/clear '),
             testDirs,
             testRootDir,
@@ -532,7 +546,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/unknown-command'),
             testDirs,
             testRootDir,
@@ -566,7 +580,7 @@ describe('useCompletion', () => {
         ] as unknown as SlashCommand[];
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/memory'), // Note: no trailing space
             testDirs,
             testRootDir,
@@ -604,7 +618,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/memory'),
             testDirs,
             testRootDir,
@@ -640,7 +654,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/memory a'),
             testDirs,
             testRootDir,
@@ -672,7 +686,7 @@ describe('useCompletion', () => {
           },
         ] as unknown as SlashCommand[];
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/memory dothisnow'),
             testDirs,
             testRootDir,
@@ -715,7 +729,7 @@ describe('useCompletion', () => {
         ] as unknown as SlashCommand[];
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/chat resume my-ch'),
             testDirs,
             testRootDir,
@@ -759,7 +773,7 @@ describe('useCompletion', () => {
         ] as unknown as SlashCommand[];
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/chat resume '),
             testDirs,
             testRootDir,
@@ -794,12 +808,14 @@ describe('useCompletion', () => {
         ] as unknown as SlashCommand[];
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('/chat resume '),
             testDirs,
             testRootDir,
             slashCommands,
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -822,12 +838,14 @@ describe('useCompletion', () => {
         await createTestFile('', 'README.md');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@s'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -856,12 +874,14 @@ describe('useCompletion', () => {
         await createTestFile('', 'src', 'index.ts');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@src/comp'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -882,12 +902,14 @@ describe('useCompletion', () => {
         await createTestFile('', 'src', 'index.ts');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@.'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -914,12 +936,14 @@ describe('useCompletion', () => {
         await createEmptyDir('dist');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@d'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfigNoRecursive,
           ),
         );
@@ -940,7 +964,7 @@ describe('useCompletion', () => {
         await createTestFile('', 'README.md');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@'),
             testDirs,
             testRootDir,
@@ -975,12 +999,14 @@ describe('useCompletion', () => {
           .mockImplementation(() => {});
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -1006,12 +1032,14 @@ describe('useCompletion', () => {
         await createEmptyDir('data');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@d'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -1040,12 +1068,14 @@ describe('useCompletion', () => {
         await createTestFile('', 'README.md');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -1073,12 +1103,14 @@ describe('useCompletion', () => {
         await createTestFile('', 'temp', 'temp.log');
 
         const { result } = renderHook(() =>
-          useCompletion(
+          useSlashCompletion(
             useTextBufferForTest('@t'),
             testDirs,
             testRootDir,
             [],
             mockCommandContext,
+            false,
+
             mockConfig,
           ),
         );
@@ -1116,12 +1148,14 @@ describe('useCompletion', () => {
 
       const { result } = renderHook(() => {
         const textBuffer = useTextBufferForTest('/mem');
-        const completion = useCompletion(
+        const completion = useSlashCompletion(
           textBuffer,
           testDirs,
           testRootDir,
           slashCommands,
           mockCommandContext,
+          false,
+
           mockConfig,
         );
         return { ...completion, textBuffer };
@@ -1158,12 +1192,14 @@ describe('useCompletion', () => {
 
       const { result } = renderHook(() => {
         const textBuffer = useTextBufferForTest('/memory');
-        const completion = useCompletion(
+        const completion = useSlashCompletion(
           textBuffer,
           testDirs,
           testRootDir,
           slashCommands,
           mockCommandContext,
+          false,
+
           mockConfig,
         );
         return { ...completion, textBuffer };
@@ -1202,12 +1238,14 @@ describe('useCompletion', () => {
 
       const { result } = renderHook(() => {
         const textBuffer = useTextBufferForTest('/?');
-        const completion = useCompletion(
+        const completion = useSlashCompletion(
           textBuffer,
           testDirs,
           testRootDir,
           slashCommands,
           mockCommandContext,
+          false,
+
           mockConfig,
         );
         return { ...completion, textBuffer };
@@ -1229,12 +1267,13 @@ describe('useCompletion', () => {
     it('should complete a file path', () => {
       const { result } = renderHook(() => {
         const textBuffer = useTextBufferForTest('@src/fi');
-        const completion = useCompletion(
+        const completion = useSlashCompletion(
           textBuffer,
           testDirs,
           testRootDir,
           [],
           mockCommandContext,
+          false,
           mockConfig,
         );
         return { ...completion, textBuffer };
@@ -1258,12 +1297,13 @@ describe('useCompletion', () => {
 
       const { result } = renderHook(() => {
         const textBuffer = useTextBufferForTest(text, cursorOffset);
-        const completion = useCompletion(
+        const completion = useSlashCompletion(
           textBuffer,
           testDirs,
           testRootDir,
           [],
           mockCommandContext,
+          false,
           mockConfig,
         );
         return { ...completion, textBuffer };
@@ -1286,12 +1326,13 @@ describe('useCompletion', () => {
 
       const { result } = renderHook(() => {
         const textBuffer = useTextBufferForTest(text);
-        const completion = useCompletion(
+        const completion = useSlashCompletion(
           textBuffer,
           testDirs,
           testRootDir,
           [],
           mockCommandContext,
+          false,
           mockConfig,
         );
         return { ...completion, textBuffer };
